+++ packages/unscharferelation/src/Unscharferelation.ts
@@ -40,7 +40,7 @@ export class Unscharferelation<P> implements IUnscharferelation<P> {
     });
 
     return Unscharferelation.of((epoque: Epoque<Array<P>>) => {
-      return Promise.all(promises).then<unknown, unknown>((heisenbergs: Array<Heisenberg<P>>) => {
+      return Promise.all(promises).then((heisenbergs: Array<Heisenberg<P>>) => {
         const arr: Array<P> = [];
         let absent: boolean = false;